RedEyesDevelopment Project: Example Author: Gvo3d Version: 0.0.0.7
 Features:
Maded first version of DAO+Service, maded spring+JUnit test of Dao and Service.

+++ src/main/java/springtestapp/dao/ContactDAO.java
@@ -17,5 +17,5 @@ public interface ContactDAO {
 
     public Contact findById(Long id);
 
-    public void delete(Integer id);
+    public void delete(Long id);
 }
